Move impl and test files for data structures
