+++ src/core/data-structures/observable.js
