+++ src/core/data-structures/promise.js
